pyproject.toml
[tool.poetry]
name = "nonebot-plugin-dcqq-relay"
version = "1.0.0"
description = "在QQ群与 Discord 之间同步消息的 nonebot2 插件"
authors = ["Autuamn_End <autuamn@qq.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = {extras = ["aiohttp", "fastapi"], version = "^2.2.1"}
nonebot-adapter-onebot = "^2.4.3"
nonebot-adapter-discord = "^0.1.7"
nonebot-plugin-orm = {extras = ["default"], version = "^0.7.3"}
pillow = "^10.3.0"
filetype = "^1.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
black = "^24.3.0"
nonemoji = "^0.1.2"
pre-commit = "^3.1.0"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = ["F", "W", "E", "UP", "ASYNC", "B", "C4", "T10", "T20", "PYI", "PT", "Q", "SIM", "TID", "RUF",]
ignore = ["E402", "B008", "B030", "RUF001", "RUF002", "RUF003",]

[tool.ruff.lint.isort]
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]
combine-as-imports = true
order-by-type = true
relative-imports-order = "closest-to-furthest"
section-order = [
    "future",
    "standard-library",
    "first-party",
    "third-party",
    "local-folder",
]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
typeCheckingMode = "standard"
reportPrivateImportUsage = false
reportShadowedImports = false
disableBytesTypePromotions = true
pythonPlatform = "All"
pythonVersion = "3.9"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
